+++ src/components/query-editor.tsx
@@ -2,7 +2,7 @@ import { css } from 'emotion';
 import React, { ChangeEvent, PureComponent } from 'react';
 import { QueryEditorProps, SelectableValue } from '@grafana/data';
 import { Button, InlineFormLabel, LegacyForms, Select, TextArea } from '@grafana/ui';
-import { DataSource } from '../DataSource';
+import { DataSource } from '../data-source';
 import {
   Aggregations,
   CommandParameters,
@@ -225,14 +225,7 @@ export class QueryEditor extends PureComponent<Props> {
             )}
 
             {CommandParameters.field.includes(command) && (
-              <FormField
-                labelWidth={8}
-                inputWidth={30}
-                value={field}
-                onChange={this.onFieldChange}
-                label="Field"
-                tooltip="Field"
-              />
+              <FormField labelWidth={8} inputWidth={30} value={field} onChange={this.onFieldChange} label="Field" />
             )}
 
             {CommandParameters.legend.includes(command) && (